% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdmm.R
\docType{package}
\name{gsdmm-package}
\alias{gsdmm-package}
\alias{_PACKAGE}
\title{gsdmm: GSDMM Short Text Clustering via Dirichlet Mixture Models}
\description{
This package implements a Dirichlet Mixture Model and accompanying Gibbs sampler for short text clustering proposed by Yin and Wang 2014.
}
\author{
\strong{Maintainer}: Till Tietz \email{ttietz2014@gmail.com}

Other contributors:
\itemize{
  \item Akiru Kato [contributor]
}

}
